+++ threedi_edits/globals.py
@@ -9,5 +9,5 @@ Created on Thu Nov 24 14:20:09 2022
 __version__ = "1.1"
 
 # Globals
-SUPPORTED_THREEDI_VERSIONS = ["0209"]
+SUPPORTED_THREEDI_VERSIONS = ["0209", "0210"]
 INSTALL_REQUIRES = ["scipy", "cached_property", "shapely==2.0a1", "threedigrid_builder"]