+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog of threedi-edits
 ===================================================
 
+1.2.1 (2022-11-15)
+----------------
+
+- Versioning support for schema 0210.
+- FID column fix.
+- Updated sideview example.
+
+
 1.0 (2022-11-15)
 ----------------